Rename mood to experience
Sometimes I feel engaged in an activity, but other times an aspect of
an activity may be engaging. Subject of the prior is me; subject of the
latter is not.

Experience is more generalized and captures both scenarios better.

Also removed Top Experience from sidebar since experience is too
free-formed to aggregate. Instead, we are showing mean / median times
and reps.