Set correct arg placeholder when editing rules
This can be done in one of two ways:
1. Render some default placeholder in rails, then update to correct
placeholder in js
2. Render correct placeholder in rails

We already have placeholder update code in js, so the first way seemed
easier. But it proved to be difficult because of TurboLinks.

So we went with the second way, which required refactored the js case
statement for determining the right placeholder for each command into a
ruby lookup table. We still nedeed this lookup table in js, so we are
loading it into a global var in application.js