Split _structured_data partial into 2 copies
One copy for activities#index and the other for messages#show (ie parse
page).

Ideally it would be nice to keep them as one, such that if a new field
is added to activity, we only have to update one partial. But the logic
for this template is starting to diverge so much that it no longer
makes sense to keep them as one file.

Parse page only logic:
- display all fields
- display 'new' label next to new types and friends

Activity index only logic:
- display only fields that aren't blank
- link types and friends as filters